+++ src/main/java/net/sf/jabref/gui/ContentSelectorDialog2.java
@@ -46,7 +46,6 @@ import javax.swing.event.ListSelectionListener;
 import net.sf.jabref.Globals;
 import net.sf.jabref.MetaData;
 import net.sf.jabref.gui.util.FocusRequester;
-import net.sf.jabref.gui.util.GUIGlobals;
 import net.sf.jabref.logic.l10n.Localization;
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
