+++ src/main/java/net/sf/jabref/gui/ImportInspectionDialog.java
@@ -87,7 +87,6 @@ import net.sf.jabref.logic.labelPattern.LabelPatternUtil;
 import net.sf.jabref.gui.undo.NamedCompound;
 import net.sf.jabref.gui.undo.UndoableInsertEntry;
 import net.sf.jabref.gui.undo.UndoableRemoveEntry;
-import net.sf.jabref.gui.util.GUIGlobals;
 import net.sf.jabref.gui.util.IconComparator;
 import net.sf.jabref.gui.util.PositionWindow;
 import net.sf.jabref.gui.desktop.JabRefDesktop;
