+++ src/main/java/net/sf/jabref/gui/KeyBindingsDialog.java
@@ -15,7 +15,6 @@
 */
 package net.sf.jabref.gui;
 
-import net.sf.jabref.gui.util.GUIGlobals;
 import net.sf.jabref.logic.l10n.Localization;
 import net.sf.jabref.util.Util;
