+++ src/main/java/net/sf/jabref/gui/MainTable.java
@@ -40,7 +40,6 @@ import net.sf.jabref.gui.renderer.CompleteRenderer;
 import net.sf.jabref.gui.renderer.GeneralRenderer;
 import net.sf.jabref.gui.renderer.IncompleteRenderer;
 import net.sf.jabref.gui.util.FirstColumnComparator;
-import net.sf.jabref.gui.util.GUIGlobals;
 import net.sf.jabref.gui.util.IconComparator;
 import net.sf.jabref.gui.util.IsMarkedComparator;
 import net.sf.jabref.gui.util.RankingFieldComparator;
