+++ src/main/java/net/sf/jabref/gui/MainTableFormat.java
@@ -23,7 +23,6 @@ import net.sf.jabref.model.entry.BibtexEntry;
 import net.sf.jabref.Globals;
 import net.sf.jabref.JabRefPreferences;
 import net.sf.jabref.model.entry.Util;
-import net.sf.jabref.gui.util.GUIGlobals;
 import net.sf.jabref.specialfields.Priority;
 import net.sf.jabref.specialfields.Rank;
 import net.sf.jabref.specialfields.ReadStatus;
