+++ src/main/java/net/sf/jabref/gui/MainTableSelectionListener.java
@@ -31,7 +31,6 @@ import javax.swing.Timer;
 
 import net.sf.jabref.gui.entryeditor.EntryEditor;
 import net.sf.jabref.gui.util.FocusRequester;
-import net.sf.jabref.gui.util.GUIGlobals;
 import net.sf.jabref.logic.l10n.Localization;
 import net.sf.jabref.logic.util.OS;
 import net.sf.jabref.model.entry.BibtexEntry;
