+++ src/main/java/net/sf/jabref/gui/SearchResultsDialog.java
@@ -43,7 +43,6 @@ import javax.swing.SwingUtilities;
 import javax.swing.table.TableColumnModel;
 
 import net.sf.jabref.gui.renderer.GeneralRenderer;
-import net.sf.jabref.gui.util.GUIGlobals;
 import net.sf.jabref.gui.util.IconComparator;
 import net.sf.jabref.model.entry.BibtexEntry;
 import net.sf.jabref.bibtex.comparator.EntryComparator;
