+++ src/main/java/net/sf/jabref/gui/fieldeditors/TextArea.java
@@ -17,9 +17,10 @@ package net.sf.jabref.gui.fieldeditors;
 
 import net.sf.jabref.Globals;
 import net.sf.jabref.gui.AutoCompleteListener;
+import net.sf.jabref.gui.GUIGlobals;
 import net.sf.jabref.gui.fieldeditors.contextmenu.FieldTextMenu;
 import net.sf.jabref.model.entry.Util;
-import net.sf.jabref.gui.util.GUIGlobals;
+import net.sf.jabref.logic.util.strings.StringUtil;
 import java.awt.*;
 
 import javax.swing.*;
