+++ src/main/java/net/sf/jabref/gui/fieldeditors/TextField.java
@@ -17,12 +17,13 @@ package net.sf.jabref.gui.fieldeditors;
 
 import net.sf.jabref.Globals;
 import net.sf.jabref.gui.AutoCompleteListener;
+import net.sf.jabref.gui.GUIGlobals;
 import net.sf.jabref.gui.actions.Actions;
 import net.sf.jabref.gui.actions.PasteAction;
 import net.sf.jabref.gui.fieldeditors.contextmenu.FieldTextMenu;
 import net.sf.jabref.gui.keyboard.KeyBinds;
 import net.sf.jabref.model.entry.Util;
-import net.sf.jabref.gui.util.GUIGlobals;
+
 import java.awt.Color;
 import java.awt.event.ActionEvent;
