+++ src/main/java/net/sf/jabref/gui/help/HelpContent.java
@@ -26,7 +26,7 @@ import javax.swing.event.HyperlinkListener;
 
 import net.sf.jabref.JabRef;
 import net.sf.jabref.JabRefPreferences;
-import net.sf.jabref.gui.util.GUIGlobals;
+import net.sf.jabref.gui.GUIGlobals;
 
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
