+++ src/main/java/net/sf/jabref/gui/preftabs/AppearancePrefsTab.java
@@ -28,7 +28,7 @@ import net.sf.jabref.*;
 import com.jgoodies.forms.builder.DefaultFormBuilder;
 import com.jgoodies.forms.layout.FormLayout;
 import net.sf.jabref.gui.ColorSetupPanel;
-import net.sf.jabref.gui.util.GUIGlobals;
+import net.sf.jabref.gui.GUIGlobals;
 import net.sf.jabref.logic.l10n.Localization;
 
 class AppearancePrefsTab extends JPanel implements PrefsTab {
