+++ src/main/java/net/sf/jabref/importer/fetcher/MedlineFetcher.java
@@ -27,7 +27,7 @@ import java.util.regex.Pattern;
 import javax.swing.JOptionPane;
 import javax.swing.JPanel;
 
-import net.sf.jabref.gui.util.GUIGlobals;
+import net.sf.jabref.gui.GUIGlobals;
 import net.sf.jabref.importer.ImportInspector;
 import net.sf.jabref.importer.fileformat.MedlineImporter;
 import net.sf.jabref.importer.OutputPrinter;
